Vendor application fix

+++ pages/dashboard/vendors/applications/index.vue
@@ -38,58 +38,109 @@
       <v-card-title class="mb-8" style="color: white; background-color: #a61c00; width: 50%; text-align: center; position: absolute; left: 10px; top: -20px; border-radius: 10px;">Your Vendor Applications</v-card-title>
       <v-btn @click="addNewVendorFormLoading" class="py-6 mb-2" color="primary" style="position: absolute; right: 10px; top: -20px; width: 25%;"><v-icon>mdi-plus</v-icon>Add New Vendor Form</v-btn>
       <template v-if="loading">
-        <v-simple-table class="pt-16">
-          <thead >
-          <tr class="d-flex justify-start">
-            <th style="color: #a61c00; width: 10%; text-align: center">Application Name</th>
-            <th style="color: #a61c00; width: 10%; text-align: center">Category</th>
-            <th style="color: #a61c00; width: 15%; text-align: center">Location Name</th>
-            <th style="color: #a61c00; width: 15%; text-align: center">Location Address</th>
-            <th style="color: #a61c00; width: 10%; text-align: center">#Questions</th>
-            <th style="color: #a61c00; width: 24%; text-align: center">Application Status</th>
-            <th style="color: #a61c00; width: 10%;">Actions</th>
-          </tr>
-          </thead>
-          <tbody>
-          <tr  v-for="(location, index) in locations" style="background: none !important;">
-            <div v-for="(service, indexService) in location.services">
-              <div class="d-flex justify-start align-center hover-select" style="border-bottom: 1px solid gray; transition: 0.3s;" v-for="(userform, indexUserForm) in service.userforms">
+<!--        <v-simple-table class="pt-16">-->
+<!--          <thead >-->
+<!--          <tr class="d-flex justify-start">-->
+<!--            <th style="color: #a61c00; width: 10%; text-align: center">Application Name</th>-->
+<!--            <th style="color: #a61c00; width: 10%; text-align: center">Category</th>-->
+<!--            <th style="color: #a61c00; width: 15%; text-align: center">Location Name</th>-->
+<!--            <th style="color: #a61c00; width: 15%; text-align: center">Location Address</th>-->
+<!--            <th style="color: #a61c00; width: 10%; text-align: center">#Questions</th>-->
+<!--            <th style="color: #a61c00; width: 24%; text-align: center">Application Status</th>-->
+<!--            <th style="color: #a61c00; width: 10%;">Actions</th>-->
+<!--          </tr>-->
+<!--          </thead>-->
+<!--          <tbody>-->
+<!--          <tr  v-for="(location, index) in locations" style="background: none !important;">-->
+<!--            <div v-for="(service, indexService) in location.services">-->
+<!--              <div class="d-flex justify-start align-center hover-select" style="border-bottom: 1px solid gray; transition: 0.3s;" v-for="(userform, indexUserForm) in service.userforms">-->
 
-                <td style="width: 10%; text-align: center" class="py-1">{{userform.name}}</td>
-                <td style="width: 10%; text-align: center" class="py-1">{{service.name}}</td>
-                <td style="width: 15%; text-align: center" class="py-1">{{location.name}}</td>
-                <td style="width: 15%; text-align: center" class="py-1">
-                  <div class="d-flex flex-column align-center">
-                    <p>{{location.address}}</p>
-                    <p>{{location.city}}, {{location.state}}</p>
-                  </div>
-                </td>
-                <td style="width: 9%; text-align: center" class="py-1">{{userform.formfields.length}}</td>
-                <td style="width: 22%;" class="py-1 center mr-10 d-flex">
-                  <v-select
-                    v-model="userform.applicationStatus"
-                    :placeholder="userform.applicationStatus"
-                    :items="applicationOptions"
-                    @change="userformEditActive(userform)"
-                  >
-                  </v-select>
-<!--                  <v-checkbox-->
-<!--                    v-if="userform.applicationStatus === 'Published - Private'"-->
-<!--                    :label="'Publish Link Publicly?'"-->
-<!--                    v-model="userform.applicationStatusLinkPublish"-->
-<!--                    class="ml-3"-->
-<!--                    @change="userformEditApplicationPublish(userform)"-->
-<!--                  ></v-checkbox>-->
-                </td>
-                <td style="width: 10%;" class="d-flex flex-column align-center">
-                  <v-btn class="my-1" color="#707070" :to="'/dashboard/vendors/applications/' + userform.id" style="color: white; width: 100%;">Edit</v-btn>
-                  <v-btn @click="deleteUserForm(userform)" class="my-1" color="primary" style="width: 100%;">Delete</v-btn>
-                </td>
-              </div>
-            </div>
-          </tr>
-          </tbody>
-        </v-simple-table>
+<!--                <td style="width: 10%; text-align: center" class="py-1">{{userform.name}}</td>-->
+<!--                <td style="width: 10%; text-align: center" class="py-1">{{service.name}}</td>-->
+<!--                <td style="width: 15%; text-align: center" class="py-1">{{location.name}}</td>-->
+<!--                <td style="width: 15%; text-align: center" class="py-1">-->
+<!--                  <div class="d-flex flex-column align-center">-->
+<!--                    <p>{{location.address}}</p>-->
+<!--                    <p>{{location.city}}, {{location.state}}</p>-->
+<!--                  </div>-->
+<!--                </td>-->
+<!--                <td style="width: 9%; text-align: center" class="py-1">{{userform.formfields.length}}</td>-->
+<!--                <td style="width: 22%;" class="py-1 center mr-10 d-flex">-->
+<!--                  <v-select-->
+<!--                    v-model="userform.applicationStatus"-->
+<!--                    :placeholder="userform.applicationStatus"-->
+<!--                    :items="applicationOptions"-->
+<!--                    @change="userformEditActive(userform)"-->
+<!--                  >-->
+<!--                  </v-select>-->
+<!--&lt;!&ndash;                  <v-checkbox&ndash;&gt;-->
+<!--&lt;!&ndash;                    v-if="userform.applicationStatus === 'Published - Private'"&ndash;&gt;-->
+<!--&lt;!&ndash;                    :label="'Publish Link Publicly?'"&ndash;&gt;-->
+<!--&lt;!&ndash;                    v-model="userform.applicationStatusLinkPublish"&ndash;&gt;-->
+<!--&lt;!&ndash;                    class="ml-3"&ndash;&gt;-->
+<!--&lt;!&ndash;                    @change="userformEditApplicationPublish(userform)"&ndash;&gt;-->
+<!--&lt;!&ndash;                  ></v-checkbox>&ndash;&gt;-->
+<!--                </td>-->
+<!--                <td style="width: 10%;" class="d-flex flex-column align-center">-->
+<!--                  <v-btn class="my-1" color="#707070" :to="'/dashboard/vendors/applications/' + userform.id" style="color: white; width: 100%;">Edit</v-btn>-->
+<!--                  <v-btn @click="deleteUserForm(userform)" class="my-1" color="primary" style="width: 100%;">Delete</v-btn>-->
+<!--                </td>-->
+<!--              </div>-->
+<!--            </div>-->
+<!--          </tr>-->
+<!--          </tbody>-->
+<!--        </v-simple-table>-->
+
+        <v-data-table
+          :headers="headersApplicationTemplateVal"
+          :items="applicationTemplateVal"
+          :items-per-page="10"
+          class="pt-16"
+          :expanded.sync="expanded"
+          show-expand
+          single-expand
+        >
+          <template v-slot:expanded-item="{ headers, item }">
+            <td :colspan="headers.length">
+              <v-simple-table>
+                <template v-slot:default>
+                  <thead>
+                  <tr>
+                    <th>Question</th>
+                    <th>Name</th>
+                  </tr>
+                  </thead>
+                  <tbody>
+                  <tr v-for="app in item.formfields" :key="app.id">
+                    <td>Question# {{(app.order + 1)}}</td>
+                    <td>{{ app.name }}</td>
+                  </tr>
+                  </tbody>
+                </template>
+              </v-simple-table>
+            </td>
+          </template>
+
+          <template v-slot:item.formfields="{ item }">
+            <p v-if="item.formfields.length">{{item.formfields.length}}</p>
+            <p v-else>0</p>
+          </template>
+
+          <template v-slot:item.applicationStatus="{ item }">
+            <v-select
+              v-model="item.applicationStatus"
+              :placeholder="item.applicationStatus"
+              :items="applicationOptions"
+              @change="userformEditActive(item)"
+            >
+            </v-select>
+          </template>
+
+          <template v-slot:item.actions="{item}">
+            <v-btn class="my-1" color="#707070" :to="'/dashboard/vendors/applications/' + item.id" style="color: white; width: 100%;">Edit</v-btn>
+            <v-btn @click="deleteUserForm(item)" class="my-1" color="primary" style="width: 100%;">Delete</v-btn>
+          </template>
+        </v-data-table>
       </template>
     </v-card>
     </transition>
@@ -507,6 +558,7 @@ import draggable from "vuedraggable"
         locations: [],
         services: [],
         userForms: [],
+        applicationTemplateVal: [],
         companyDocuments: [],
         finishedFormFields: false,
         totalLength: 0,
@@ -539,6 +591,15 @@ import draggable from "vuedraggable"
         addLocation: {},
         applicationTemplates: [],
         companyTemplates: [],
+        headersApplicationTemplateVal: [
+          { text: 'Application Name', value: 'name', class: 'primary--text font-weight-regular' },
+          { text: 'Category', value: 'serviceName', class: 'primary--text font-weight-regular' },
+          { text: 'Location Name', value: 'locationName', class: 'primary--text font-weight-regular' },
+          { text: 'Location Address', value: 'locationAddress', class: 'primary--text font-weight-regular' },
+          { text: '#Questions', value: 'formfields', class: 'primary--text font-weight-regular' },
+          { text: 'Application Status', value: 'applicationStatus', class: 'primary--text font-weight-regular' },
+          { text: 'Actions', value: 'actions', sortable: false, class: 'primary--text font-weight-regular' },
+        ],
         headers: [
           {
             text: 'ID',
@@ -705,8 +766,9 @@ import draggable from "vuedraggable"
         setTimeout(() => {
           console.log(this.userForms, 'userForms with formfields Locations');
           console.log(this.locations, 'locations Locations');
+          console.log(this.applicationTemplateVal, 'APPLICATION TEMPLATE FOR LOOP TABLE')
           this.loading = true;
-        }, 500)
+        }, 2000)
 
       },
       async getServices(id) {
@@ -738,12 +800,15 @@ import draggable from "vuedraggable"
         if(this.locations[valueServices].services[valueUserForms] !== 'There are no services') {
           await this.$http.get('https://www.sowerkbackend.com/api/userforms/byServiceId/' + id)
             .then(async (response) => {
-              this.totalLength += response.data.length;
-              console.log(this.totalLength, 'totalLength!!!!!!!!!!!!');
-              console.log(response.data, 'userforms response.data');
-              console.log(valueServices, 'valueServices', valueUserForms, 'valueUserForms');
-              console.log(this.locations[valueServices], 'locationsValueServices', this.locations[valueServices].services[valueUserForms])
-              this.locations[valueServices].services[valueUserForms].userforms = response.data;
+              // this.totalLength += response.data.length;
+              // console.log(this.totalLength, 'totalLength!!!!!!!!!!!!');
+              // console.log(response.data, 'userforms response.data');
+              // console.log(valueServices, 'valueServices', valueUserForms, 'valueUserForms');
+              // console.log(this.locations[valueServices], 'locationsValueServices', this.locations[valueServices].services[valueUserForms])
+              // this.locations[valueServices].services[valueUserForms].userforms = response.data;
+              setTimeout(() => {
+
+              }, 250);
               for(let i=0; i<response.data.length; i++) {
                 let userForm = {
                   applicationStatus: response.data[i].applicationStatus,
@@ -752,17 +817,32 @@ import draggable from "vuedraggable"
                   name: response.data[i].name,
                   service_id: response.data[i].service_id,
                   formfields: []
-                }
+                };
+                let userForm2 = {
+                  applicationStatus: response.data[i].applicationStatus,
+                  applicationStatusLinkPublish: response.data[i].applicationStatusLinkPublish,
+                  id: response.data[i].id,
+                  name: response.data[i].name,
+                  service_id: response.data[i].service_id,
+                  formfields: [],
+                  locationName: this.locations[valueServices].name,
+                  locationAddress: this.locations[valueServices].address + " " + this.locations[valueServices].city + ", " + this.locations[valueServices].state + " " + this.locations[valueServices].zipcode,
+                  serviceName: this.locations[valueServices].services[valueUserForms].name
+                };
                 if(userForm.applicationStatus === 0) {
                   userForm.applicationStatus = 'Unpublished'
+                  userForm2.applicationStatus = 'Unpublished'
                 } else if (userForm.applicationStatus === 1) {
                   userForm.applicationStatus = 'Published - Public'
+                  userForm2.applicationStatus = 'Published - Public'
                 } else {
                   userForm.applicationStatus = 'Published - Private'
+                  userForm2.applicationStatus = 'Published - Private'
                 }
                 await this.userForms.push(userForm);
+                await this.applicationTemplateVal.push(userForm2);
                 // console.log(this.valueServices, 'this.valueServices', this.valueUserForms, 'this.valueUserForms', this.locations, 'this.locations')
-                this.locations[valueServices].services[valueUserForms].userforms[i] = userForm;
+                // this.locations[valueServices].services[valueUserForms].userforms[i] = userForm;
                 console.log(this.locations[valueServices].services[valueUserForms].userforms[i], 'userform');
                 await this.getFormFields(response.data[i].id);
               }
@@ -779,8 +859,9 @@ import draggable from "vuedraggable"
           .then(async (response) => {
             console.log(response.data, 'formfields for userform', id);
             this.userForms[this.valueFormFields].formfields = response.data;
-            this.locations[this.valueServices].services[this.valueUserForms].userforms[this.valueFormFields]["formfields"] = response.data;
-            console.log(this.locations[this.valueServices].services[this.valueUserForms].userforms[this.valueFormFields].formfields, 'userforms with formfields')
+            this.applicationTemplateVal[this.valueFormFields].formfields = response.data;
+            // this.locations[this.valueServices].services[this.valueUserForms].userforms[this.valueFormFields]["formfields"] = response.data;
+            // console.log(this.locations[this.valueServices].services[this.valueUserForms].userforms[this.valueFormFields].formfields, 'userforms with formfields')
             this.valueFormFields++
             // console.log(this.valueFormFields, 'valueFormFields');
           })
@@ -1017,14 +1098,16 @@ import draggable from "vuedraggable"
       },
       async deleteUserForm(userform) {
         console.log(userform, 'userform');
-        for(let i=0; i<userform.formfields.length; i++) {
-          await this.$http.delete('https://www.sowerkbackend.com/api/formfields/' + userform.formfields[i].id)
-            .then(response => {
-              console.log(response, 'success in deleting formfields');
-            })
-            .catch(err => {
-              console.log(err, 'err in deleting formfields')
-            })
+        if(userform.formfields.length > 0) {
+          for(let i=0; i<userform.formfields.length; i++) {
+            await this.$http.delete('https://www.sowerkbackend.com/api/formfields/' + userform.formfields[i].id)
+              .then(response => {
+                console.log(response, 'success in deleting formfields');
+              })
+              .catch(err => {
+                console.log(err, 'err in deleting formfields')
+              })
+          }
         }
         await this.$http.delete('https://www.sowerkbackend.com/api/userforms/' + userform.id)
           .then(response => {
